requirements.txt
datasets == 2.21.0
langchain_community==0.2.12
torch == 2.4.1
torchaudio==2.4.1
transformers==4.44.0
sentence_transformers==2.7.0
accelerate == 1.0.1
librosa==0.10.2
langchain_core==0.2.41
flask
langchain_huggingface==0.0.2
langchain_chroma==0.1.3
langchain_cohere==0.2.2
langchain_groq==0.1.10
llmsherpa==0.1.4
python-docx
assemblyai